Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt' Ensured that all projects compile Removed compiler warnings Better setting of server_version variable. Fix that make_win_src_distribution creates the privilege tables.
VC++Files/bdb/bdb.dsp:
  Small, automatic changes
VC++Files/client/mysql.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/client/mysqladmin.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/client/mysqlclient.dsp:
  Removed files that should only be used with mysql command line client
VC++Files/client/mysqldump.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/client/mysqlimport.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/client/mysqlshow.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/comp_err/comp_err.dsp:
  Automatic changes
VC++Files/dbug/dbug.dsp:
  Automatic changes
VC++Files/heap/heap.dsp:
  automatic changes
VC++Files/innobase/innobase.dsp:
  Automatic changes
VC++Files/isam/isam.dsp:
  Automatic changes
VC++Files/isamchk/isamchk.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/libmysql/libmysql.dsp:
  Automatic changes
VC++Files/libmysqld/examples/test_libmysqld.dsp:
  Add missing files
VC++Files/libmysqld/libmysqld.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/libmysqltest/myTest.dsp:
  Automatic changes
VC++Files/merge/merge.dsp:
  Automatic changes
VC++Files/my_print_defaults/my_print_defaults.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/myisam/myisam.dsp:
  automatic changes
VC++Files/myisam_ftdump/myisam_ftdump.dsp:
  automatic changes
VC++Files/myisamchk/myisamchk.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/myisamlog/myisamlog.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/myisammrg/myisammrg.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/myisampack/myisampack.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/mysql.dsw:
  Automatic changes
VC++Files/mysqlbinlog/mysqlbinlog.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/mysqlcheck/mysqlcheck.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/mysqldemb/mysqldemb.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/mysqlserver/mysqlserver.dsp:
  Automatic changes
VC++Files/mysqlshutdown/mysqlshutdown.dsp:
  Automatic changes
VC++Files/mysqlwatch/mysqlwatch.dsp:
  Automatic changes
VC++Files/mysys/mysys.dsp:
  Automatic changes
VC++Files/pack_isam/pack_isam.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/perror/perror.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/regex/regex.dsp:
  Automatic changes
VC++Files/replace/replace.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/sql/mysqld.dsp:
  Added support for projects 'classic', 'classic nt', 'pro' and 'pro nt'
VC++Files/strings/strings.dsp:
  Removed duplicate code for strnlen
VC++Files/test1/test1.dsp:
  Automatic changes
VC++Files/thr_test/thr_test.dsp:
  Automatic changes
VC++Files/vio/vio.dsp:
  Automatic changes
VC++Files/zlib/contrib/asm386/zlibvc.dsp:
  Automatic changes
VC++Files/zlib/zlib.dsp:
  Automatic changes
extra/my_print_defaults.c:
  Fixed bug in --verbose
include/m_string.h:
  Portability fix
include/mysql_embed.h:
  Better setting of server_version variable
include/mysql_version.h.in:
  Better license text handling
innobase/pars/pars0lex.l:
  Remove compiler warnings
innobase/trx/trx0sys.c:
  Remove compiler warnings
libmysqld/lib_sql.cc:
  Better setting of server_version variable
libmysqld/libmysqld.def:
  Add functions needed for mysql command line client
myisam/myisam_ftdump.c:
  Remove compiler warnings
mysys/sha1.c:
  Remove compiler warnings
scripts/make_win_src_distribution.sh:
  Safety fix
scripts/mysql_install_db.sh:
  Backport from 4.1 to allow make_win_src_distribution create the privilege tables
sql/Makefile.am:
  Add new file mysqld_suffix.h
  Remove not used file sql_olap.h
sql/ha_innodb.cc:
  Remove not used variable
sql/mysqld.cc:
  Better setting of server_version variable
sql/set_var.cc:
  Fixed bug when showing lower_case_file_system
strings/ctype-tis620.c:
  Remove compiler warnings

+++ myisam/myisam_ftdump.c
@@ -81,7 +81,7 @@ int main(int argc,char *argv[])
 
   {
     char *end;
-    inx= strtoll(argv[1], &end, 10);
+    inx= (uint) strtoll(argv[1], &end, 10);
     if (*end)
       usage();
   }
